Fix incorrect log format strings
Fix a number of instances of incorrect format strings, and missing
parameters, in the logging code.

The core logging function is now adorned with the proper GCC function
attributes, preventing future log formatting issues of this nature.

None of these bugs are likely to have caused any crashes or other
issues (other than malformed log messages) in any released version of
XCM.

+++ libxcm/tp/tls/xcm_tp_btls.c
@@ -1034,7 +1034,7 @@ static int btls_send(struct xcm_socket *__restrict s,
     UT_RESTORE_ERRNO(write_errno);
 
     if (rc > 0) {
-	LOG_SEND_ACCEPTED(s, buf, rc);
+	LOG_SEND_ACCEPTED(s, buf, (size_t)rc);
 	XCM_TP_CNT_BYTES_INC(bts->conn.cnts, from_app, rc);
 
 	LOG_LOWER_DELIVERED_PART(s, rc);
